+++ Sources/KikiCommerceCore/KikiProCompatibility.swift
@@ -1,0 +1,26 @@
+@available(*, deprecated, renamed: "KikiAccessManager")
+public typealias KikiProAccessManager = KikiAccessManager
+
+@available(*, deprecated, renamed: "KikiAccessPlan")
+public typealias KikiProPlan = KikiAccessPlan
+
+@available(*, deprecated, renamed: "KikiAccessPlanProduct")
+public typealias KikiProPlanProduct = KikiAccessPlanProduct
+
+@available(*, deprecated, renamed: "KikiAccessState")
+public typealias KikiProAccessStatus = KikiAccessState
+
+@available(*, deprecated, renamed: "KikiAccessRenewalState")
+public typealias KikiProRenewalState = KikiAccessRenewalState
+
+@available(*, deprecated, renamed: "KikiAccessDebugMode")
+public typealias KikiProAccessDebugMode = KikiAccessDebugMode
+
+@available(*, deprecated, renamed: "KikiAccessStorageKeys")
+public typealias KikiProAccessStorageKeys = KikiAccessStorageKeys
+
+@available(*, deprecated, renamed: "KikiAccessConfiguration")
+public typealias KikiProAccessConfiguration = KikiAccessConfiguration
+
+@available(*, deprecated, renamed: "KikiAccessPlanPackageMetadata")
+public typealias KikiProPlanPackageMetadata = KikiAccessPlanPackageMetadata