+++ Sources/KikiCommercePresentation/KikiProCompatibility.swift
@@ -1,0 +1,11 @@
+@available(*, deprecated, renamed: "KikiAccessPaywallSheet")
+public typealias KikiProPaywallSheet = KikiAccessPaywallSheet
+
+@available(*, deprecated, renamed: "KikiAccessPaywallCopy")
+public typealias KikiProPaywallCopy = KikiAccessPaywallCopy
+
+@available(*, deprecated, renamed: "KikiAccessPaywallLink")
+public typealias KikiProPaywallLink = KikiAccessPaywallLink
+
+@available(*, deprecated, renamed: "KikiAccessPaywallContext")
+public typealias KikiProPaywallPresentationContext = KikiAccessPaywallContext